Modularity refactoring
Added the nou module which contains game logic and discord API helper functions.
Moved the start, begin, and stop commands to the nou module in nou.go.
Added functions in nou module in discord.go that aid in communication with the Discord API.
Added callback functions to handle reaction behavior.